Ajout du support pour nombres 16 bits dans debug

+++ tp/tp9/lib/include/debug.h
@@ -24,7 +24,7 @@ public:
      * 
      * \param nombre : le nombre a transmettre
      */
-    static void out(const uint8_t& nombre);
+    static void out(const uint16_t& nombre);
 
     /**
      * Affichage nombre
@@ -32,7 +32,7 @@ public:
      * \param nombre : le nombre a transmettre
      * \param base : la base du nombre
      */
-    static void out(const uint8_t& nombre, const uint8_t& base);
+    static void out(const uint16_t& nombre, const uint8_t& base);
 
     /**
      * Affichage string
